tests: move alembic tests
